ADD: escalado. Los textos no escalan. Se puede mejorar

+++ code/0hn0/GameLogic/src/main/java/es/ucm/arblemar/gamelogic/Celda.java
@@ -30,7 +30,7 @@ public abstract class Celda extends GameObject {
     //  Margen de la celda
     protected int margen;
     //  Radio del circulo
-    protected int radio;
+    protected float radio;
     //  Valor de esta celda
     protected int valor;
