+++ code/0hn0/GameLogic/src/main/java/es/ucm/arblemar/gamelogic/CeldaRoja.java
@@ -47,10 +47,10 @@ public class CeldaRoja extends Celda {
     @Override
     public void render(es.ucm.arblemar.engine.Graphics g) {
         g.setColor(color);
-        g.fillCircle(pos,radio);
+        g.fillCircle(pos, (int)radio);
         if (valor > 0) {
             g.setColor(0XFFFFFFFF);
-            g.drawText(Integer.toString(valor), pos._x + (float)radio / 2, pos._y + (float)radio / 2);
+            g.drawText(Integer.toString(valor), (int)(pos._x + (radio / 2)), (int)(pos._y + (radio / 2)));
         }
 //        if(interactive){
 //            g.setColor(0X333333FF);
